// src/generated/api/catalog/models/productSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { ProductSummaryDtoProductType } from './productSummaryDtoProductType';
import type { ProductSummaryDtoStatus } from './productSummaryDtoStatus';
import type { ProductSummaryDtoCurrency } from './productSummaryDtoCurrency';

export interface ProductSummaryDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  /**
   * Sellable variant represented by minPrice; use this ID for quick-add cart actions
   * @nullable
   * @pattern ^[1-9][0-9]*$
   */
  defaultVariantId?: string | null;
  /**
   * Immutable SKU of defaultVariantId
   * @nullable
   */
  defaultVariantSku?: string | null;
  productNo: string;
  name: string;
  slug: string;
  brand?: string;
  primaryCategory?: string;
  productType: ProductSummaryDtoProductType;
  status: ProductSummaryDtoStatus;
  version: number;
  /** @nullable */
  minPrice?: string | null;
  currency: ProductSummaryDtoCurrency;
  /** @nullable */
  imageUrl?: string | null;
}